requirements.txt
Django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==12.0.0
boto3==1.42.4
psycopg2-binary==2.9.11
python-decouple==3.8
